Revert "ZBUG-1315: Fixes after soap test"
This reverts commit aa2eab3f5e10611b550325cb5e963b5bd8d841c9.

+++ store/src/java/com/zimbra/cs/db/DbVolumeBlobs.java
@@ -59,7 +59,7 @@ public final class DbVolumeBlobs {
             stmt.setShort(pos++, vol.getId());
             stmt.setInt(pos++, mbox.getId());
             stmt.setInt(pos++, item.getId());
-            stmt.setLong(pos++, item.getSavedSequence());
+            stmt.setInt(pos++, item.getSavedSequence());
             stmt.setString(pos++, item.getDigest());
             stmt.setBoolean(pos++, false);
             stmt.executeUpdate();
@@ -82,7 +82,7 @@ public final class DbVolumeBlobs {
             stmt.setShort(pos++, volId);
             stmt.setInt(pos++, info.mailboxId);
             stmt.setInt(pos++, info.itemId);
-            stmt.setLong(pos++, info.revision);
+            stmt.setInt(pos++, info.revision);
             stmt.setString(pos++, info.digest);
             stmt.setBoolean(pos++, false);
             stmt.executeUpdate();
